gpuav: More more stuff to descriptor_checks
